Share question form and shortcuts between native and plugin requests
