+++ packages/plugin-build/src/runtime-shims.mjs
@@ -86,6 +86,7 @@ export const RUNTIME_SLOT_BY_SPECIFIER = Object.freeze({
   "tailwind-merge": "tailwindMerge",
   "class-variance-authority": "classVarianceAuthority",
   [SHARED_UI_ICON_SPECIFIER]: "sharedUiIcon",
+  "@bb/shared-ui/question-form-host": "questionFormHost",
 });
 
 /** The npm package owning a specifier: `react/jsx-runtime` → `react`. */
