+++ packages/shared-ui/package.json
@@ -343,6 +343,21 @@
       "source": "./src/lib/portal-scope.ts",
       "types": "./src/lib/portal-scope.ts",
       "default": "./src/lib/portal-scope.ts"
+    },
+    "./question-form": {
+      "source": "./src/components/ui/question-form.tsx",
+      "types": "./src/components/ui/question-form.tsx",
+      "default": "./src/components/ui/question-form.tsx"
+    },
+    "./question-form-state": {
+      "source": "./src/components/ui/question-form-state.ts",
+      "types": "./src/components/ui/question-form-state.ts",
+      "default": "./src/components/ui/question-form-state.ts"
+    },
+    "./question-form-host": {
+      "source": "./src/components/ui/question-form-host.tsx",
+      "types": "./src/components/ui/question-form-host.tsx",
+      "default": "./src/components/ui/question-form-host.tsx"
     }
   },
   "scripts": {
